// backend/src/game/types.ts
export enum WeaponType {
  // Assault Rifles
  AK47 = 'AK47',
  M16 = 'M16',
  Galil = 'Galil',
  FAMAS = 'FAMAS',
  // SMGs
  MP40 = 'MP40',
  AK74u = 'AK74u',
  MP5 = 'MP5',
  // Shotguns
  Olympia = 'Olympia',
  SPAS12 = 'SPAS12',
  // Snipers
  L96A1 = 'L96A1',
  Dragunov = 'Dragunov',
  // Wonder Weapons
  Raygun = 'Raygun',
  Thundergun = 'Thundergun',
  WunderWaffe = 'WunderWaffe',
}

export enum WeaponRarity {
  Common = 'Common',
  Uncommon = 'Uncommon',
  Rare = 'Rare',
  Epic = 'Epic',
  Legendary = 'Legendary',
  Mythic = 'Mythic',
}